split off ApiDataAccessObject from ProviderHelper
